Store mindmap source in PNG images
This change will store mindmap source diagram as metadata in PNG images or as
comment in SVG images. It is very useful when only image, not source, is
available and the change to mindmap is required.

This functionality was/is already available for other types of diagrams
but was not working for mindmap.

As described in `help` to retrieve mindmap sources use:

```
java -jar plantuml.jar -metadata <mindmap-image>.png
```
To get the sources from SVG image simply open SVG file in text editor
and copy sources from `comment element` at the end of the file.