Update Colors according to Archiate off. Specs
This is mainly a proposal.
The colors at the moment are sometimes quite "loud" and on some it's a bit hard to read the (black) text on it. The colors from the Archimate Specifications (https://pubs.opengroup.org/architecture/archimate3-doc/chap03.html#_Toc10045294) are more subtle and you can better read text on it. It is not necessary to change the colors of the sprites too. But that's just my opinion.