javac --release to build for java8, with java11+
javac since java9 has a new flat --release to build a binary which can be run
with an older java. the public API of that older release must be followed as
well.

the flags used before, -source, -target do not do this. therefor it got
unnoticed that plantuml did not compile and run in java7 any more, despite
beeing compiled for for java7.

https://github.com/plantuml/plantuml/issues/898