permit "mvn test" with java8.
activate javac --release only on higher java versions. it allows to compile
with java11+, but permits afterwards running tests with java8 to make sure
rendering errors can be catched.

parmit the syntax of the source to include java17, and tell the user about
it instead of just failing the compilation.