Refactor createEntityNode() into createEntityNode() so we end up with an unclosed node
This may not be necessary since I don't think there are any constraints
to when you add a child to it's parent or parent-to-be, but it keeps
with the motif of shoving things into a stack and not popping elements
until we're done adding everything into it. But this motif also adds the
complexity of needing to read it's return value and pop the number of
elements off when done.