define -> arrows as request and <- as response
Instead of defining that every message from the left to the right side is always a request and every message from the right to the left side is always a response, I propose to used the arrow direction from the PlantUML diagram code.

Each arrow with the arrow direction on the right (`->`) is a request and each arrow with the arrow direction on the left (`<-`) is a response.

This way the users can define themselves what is a request and what is a response.