feat: add builtin function `%boolval`
This change adds the `%boolval` function to improve usability of boolean
values and conditionals. This is especially helpful when the diagram
data is generated in a previous workflow step, using a conventional
programming or scripting language.

In this first implementation, the boolean strings are defined as:

* True values: `true`, `1`
* False values: `false`, `0`

These values are case-sensitive, meaning `True` `TRUE` are also treated
as true values (likewise for false values).

This builtin function helps mitigate some user confusion with the use of
boolean values, specifically that all strings including `"0"` and
`"false"` are treated as true values.

Existing behaviour (integer `0` is false, all other numbers and strings
are true) is not affected, and can be used by simply not calling the
new `%boolval` function.

Other minor improvements:

1. Parsing error in `%intval` will now raise an exception. This improves
   tool behaviour consistency, for example, when `-noerror` flag is used
   from the CLI.
2. Functions in `TContext.java` are arranged in sorted order. This code
   readability improvement does not change the tool behaviour from the
   user's perspective.