Add floating participants header to interactive SVG sequence diagram (#1995)
* Group sequence diagram SVG elements together in  interactive mode

* Make interactive SVG scripts work for multiple images

Both the "default" and "onmouseinteractivefooter" failed to work properly
when multiple interactive SVGs were in the same HTML page.

In the case of the latter, the root cause was the external SVG library, so
the script has been refactored to standard Document APIs instead.

* Enable floating sequence diagram headers for interactive SVGs

It would have been nice to be able to use the standard CSS `position: sticky`
but this won't work inside SVGs, plus it has a tendency to stop working
when an ancestor has an `overflow` style set.