Display participant name when hovering over lifeline
As per suggestion by @ioplker, it uses the SVG `<title>` element to help users of large SVG sequence
diagrams to see which participant belongs to a given lifeline without
having to scroll up to view the header. Making the headers "float" is
a much nicer user experience (can see all participant names at once,
and no need to wait several seconds for a tooltip to appear) but
requires the special `!pragma svginteractive true` directive to be
added, whereas this `<title>` approach can be enabled by default for all SVG sequence diagrams.

The implementation includes a new `<rect>` behind each participant's
vertical line, which acts as a hover target for the mouse. (Trying to
hover over a line that's a 1px wide is not practical.)

In `SvgGraphics`, the new `<rect>` cannot have a `fill="none"`
attribute, as this makes it ineligible for mouse detection, so I have
changed this to `fill="transparent"` for all cases. I also removed a
number of explicit branches for `#00000000` fill colour value; these
branches were already redundant due to the value being changed by the
existing `fixColor()` method.