intentionally break compilation to identify usages of `getIHtmlColorSet`
**details:**

The goal is to use the compiler to detect all occurrences of `getIHtmlColorSet` so that they can potentially be replaced with the new method `HColor getColor(PName pname)`.
If a replacement is not possible or straightforward, the call should be temporarily updated to `getIHtmlColorSetFooBar29()`.

Once all usages have been reviewed and updated, `getIHtmlColorSetFooBar29()` will be renamed back to `getIHtmlColorSet` to restore the original method name.